requirements.txt
docker-py
Flask-MongoEngine
Flask-uWSGI-WebSocket